+++ src/App.js
@@ -1,54 +1,19 @@
 import React from 'react';
-import axios from 'axios';
-import Movie from './Movie';
-import './App.css';
+import Main from './routes/Main';
+import About from './routes/About';
+import Detail from './routes/Detail';
+import Navigation from './components/Navigation';
+import {HashRouter, Route} from 'react-router-dom';
+import "./App.css";
 
-class App extends React.Component{
-  state = {
-    isLoading: true,
-    movies: []
-  }
 
-  getMovies = async() => {
-    const {
-      data: {
-        data: { movies }
-      }
-    }  = await axios.get("https://yts-proxy.now.sh/list_movies.json?sort_by=rating")
-    this.setState({ movies, isLoading : false })
-  }
-
-  async componentDidMount() {
-    this.getMovies();
-  }
-
-  render() {
-    const { isLoading, movies } = this.state
-    return ( 
-      <section className = "container">
-        { isLoading 
-          ? (
-          <div className="container">
-            <span className="loader__text">Loading...</span>
-          </div>
-          ) : ( 
-            <div className="movies">
-              {movies.map(movie=>(
-                <Movie
-                  key={movie.id}
-                  id = {movie.id}
-                  year = {movie.year}
-                  title = {movie.title}
-                  summary = {movie.summary}
-                  poster = {movie.medium_cover_image}
-                  genres = {movie.genres}
-                />
-              ))}
-            </div>
-          )}
-      </section>
-    )
-  }
+function App() {
+  return <HashRouter>
+    <Navigation />
+    <Route path='/' exact={true} component={Main} />
+    <Route path='/about' component={About} />
+    <Route path='/movie/:id' component={Detail} />
+  </HashRouter>
 }
 
-export default App;
+export default App
