+++ src/routes/About.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './About.css';
+
+function About() {
+  return (
+    <div className="about__container">
+      <span>
+        React sample Clone
+      </span>
+      <span>− from NomadCoder</span>
+    </div>
+  );
+}
+export default About;
